backend/requirements.txt
Django==5.2.2
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
gunicorn==23.0.0